# Part2_Infrastructure/tools/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified execution-quality, pre-trade-risk and research gateway. Module A: cross-venue L2 TCA \u00b7 Module B: pre-trade risk & kill switch \u00b7 Module C: asynchronous parametric backtesting.",
    "title": "AlphaEngine Trading Automation \u2014 Execution Gateway & Risk Portal",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/research/openbb/bars": {
      "get": {
        "operationId": "openbb_bars_api_research_openbb_bars_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9.\\-]+$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset",
            "required": false,
            "schema": {
              "default": "equity",
              "pattern": "^(equity|crypto)$",
              "title": "Asset",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "pattern": "^(15m|1h|4h|1d)$",
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 10,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-AlphaEngine-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Alphaengine-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Openbb Bars Api Research Openbb Bars Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openbb Bars",
        "tags": [
          "C \u00b7 Research"
        ]
      }
    },
    "/api/research/openbb/fundamentals": {
      "get": {
        "operationId": "openbb_fundamentals_api_research_openbb_fundamentals_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9.\\-]+$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-AlphaEngine-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Alphaengine-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Openbb Fundamentals Api Research Openbb Fundamentals Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openbb Fundamentals",
        "tags": [
          "C \u00b7 Research"
        ]
      }
    },
...
    "/api/research/openbb/news": {
      "get": {
        "operationId": "openbb_news_api_research_openbb_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbols",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Symbols",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-AlphaEngine-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Alphaengine-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Openbb News Api Research Openbb News Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openbb News",
        "tags": [
          "C \u00b7 Research"
        ]
      }
    },
    "/api/research/openbb/quote": {
      "get": {
        "operationId": "openbb_quote_api_research_openbb_quote_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9.\\-]+$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset",
            "required": false,
            "schema": {
              "default": "equity",
              "pattern": "^(equity|crypto)$",
              "title": "Asset",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-AlphaEngine-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Alphaengine-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Openbb Quote Api Research Openbb Quote Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openbb Quote",
        "tags": [
          "C \u00b7 Research"
        ]
      }
    },
...
  }
}